add first uix unit test.
The GraphicUnitTest allow you to test the rendering of a specific widget tree.
You can setup the delay before doing the screenshot
When the screenshot is done, it will be compared to the reference
Also, an html is generated with reference image, current image, and source code involved.

Be careful, the rendering context is reseted between multiple render() call. You need to recreate your object from scratch.