Initial support for animation of gif's and images in .zip  in core/Image class
An Image widget that has a source as a .gif or a .zip
will now automatically animate.

The core Image class fires a 'on_texture_changed' event that has to be
handled by any widget that wants to display animation for the images.
The uix/Image widget class does this.

      core_image.bind(on_texture_changed = on_tex_changed)

def on_tex_changed(self)
    self.texture = self.core_image.texture

The following functions/variables allow control of animation in
coreimagee

    reset_anim(True/False)   enable or disable animation on
                             by default(if image supports it)
    anim_frame_delay = .22   time between frames in seconds

 On branch png_gif_animation