Selection had no memory of previously selected items, because the data used in the examples were just strings primarily. To further development with a more realistic environment, a DataStore class was created, with simple set and get property methods, operating on a key into a dict of records. The simple scheme for the records is that they must all contain an is_selected boolean for storing the state of selection. The fruit dataset was modified accordingly, into two datastores: categories and fruits. A new simple dataset was also created for integer data used in several examples.
For view instances to be associated with data records in a datastore, an index
property was added to SelectableItem. This index into the data list is used in
selection operations to keep selection of data items in sync with item view
instances.

When creating item view instances in ListView there is now a lookup call to
the datastore to check for selection of the associated data item record. The
data item record tested so far carries an is_selected boolean, but the code
was written to allow for an is_selected() function.