finish implementation of DPI.
The current implement live in NumericProperty. We still have a hack cause of Text core provider: we was always multiply the font_size by 1.333 for simulating DPI on standard screen (72->96).
But now, we should'nt do that. If we remove the multiplication, all the px font_size will be reduced.
In order to not break all the apps, we are doing the * 1.333 only for font_size != px, in Label and TextInput.
You can configure the DPI with --dpi=222 or KIVY_DPI=222.