StackLayout: Implement top-to-bottom and right-to-left orientations.
This allows StackLayout to organise widgets in any combination of lr,
rl, tb, bt orientations. Widgets are aligned depending on the
orientation as well. Left to right left-aligns widgets, right to left
right-aligns widgets, top to bottom aligns widgets to the top, bottom to
top aligns widgets to the bottom.