Edit `BoundedNumericProperty.check` to use "%f" rather than "%d" for float max/min in ValueError error message.
If the supplied value is outside the bounds of BoundedNumericProperty, a ValueError is raised by `BoundedNumericProperty.check`.  However, the error message inside the ValueError is incorrect if the max or min are floats, since the error message uses "%d" rather than "%f".

This means for example that if the min is 0.6 and you supply a value of 0.4, the error message will tell you that your value was below the minimum bound of "0", which is incorrect for 2 reasons.