Clarifying the documentation of ScatterLayout
The previous documentation was based in the old RelativeLayout (based in a Scatter). The documentation didn't indicate the difference between Scatter and ScatterLayout. It is also more clear if we compare the ScatterLayout against the RelativeLayout since they behaviour is the same (except for the scatter functionality). Scatter is already relative so the only difference that I can see is the possibility of using the hint properties.