Implemented sort and reverse for list ops
Sort and reverse ops have a start call where pre-sort data-to-view
associations are made, so that in the finalization step cached views can
be rearranged and selection maintained.

Added buttons in the test app for changing data between strings and
objects.

Fixed a bug with the dict setitem_add op.