Solved issue #1388 where ``lang._handlers`` keeps widgets alive.
The original problem was caused by 2 things: 1) during
``create_handler``, in the binding clause, we did not make sure that
``f`` is a weak proxy. 2) during template creation, the ``ctx``
dictionary is passed in as is, instead of making sure to use only the
weak proxies of the corresponding widgets in the dictionary.