Added several list examples, and fixed others
Renamed the example app that has been used for developing new
oberverable list and dict variants to be consistent with the other list
examples.

Added the list_of_carousels.py example, which was made to solve a
mistaken problem about namespacing index in ListView and Carousel.
Turned out that this is not a problem.

Fixed a few lingering problems in the list examples.

Edited a few comments.