Added notes about storing transform functions
Now that binding transforms are available, because lambdas are limited
to one line, and often long expressions will break pep8, we will be
writing small binding transform functions. These can be interleaved,
in-line, between controllers and views, such as ListView in the
examples, but for larger apps, these functions could be stored in
utility modules for better code organization and name-spacing.

For example, we could have utils/transform_functions.py.