args_converter now are on list item views
Kivy does not have formal data models (no Record class or the
equivalent), where database-style attributes and relationships such as
one-to-many, many-to-many, etc. are specified. The closest thing we have
to a data model is SelectableDataItem, which only adds a ksel selection
object. However, the args_converters used for list item views take on
some of the role of a formal data model class, by providing a way to
specify how individual attributes on a data item are prepared and
calculated for list item views. It makes sense then, to put the
args_converter closest to where this conversion happens, and it is
better to put it on the view than on data. So, SelectableView and
related list item view base classes now have a default args_converter,
adn the examples were changed to remove the args_converters and use the
default ones, or to make new ones that have their own args_converter
method.

CompositeListItem is more complicated, but an attempt was made at a
default args_converter method, which uses an added component_args list
property to hold the individual attribute and conversion expressions
needed for the components of the composite. This solution will need
changing to work more fluidly in kv probably.

The anticipated work to improve how the args converter system works with
CompositeListItem will likely help with finding a better way to specify
the args_converters in general. Presently, the args_converter is a
method that returns a kwargs hash. However, perhaps, in association with
exploring for good functionality in kv, we may find a class-based way to
do it. This make take on a model record type construction.