Add warnings to width/height properties
Especially new users might stumble when trying to reference width and
height properties before the layout is done. It's not immediately
obvious why a widget that should span the whole screen, for example, has
a width of 100 (default).

This fix adds a warning to the documentation to help lessen the
confusion.