Expose a property to set the scroll distance when using a mouse wheel.
By changing scroll_distance we can achieve the same effect but this will render touch dragging unusable and buggy, especially if we set a greater value. By adding a separate property that is used only for setting the scroll distance in the case of mouse wheel scrolling we can safely use together a custom scroll_wheel_distance and touch dragging.