Placement of import config critical to opening window size.
It should be important to add that config ought to be imported right after 

"from kivy.app import App"

if the user doesn't wish to observe a blank screen with the kivy icon before it resizes. 

This issue seems to be caused by placement of import config and Config.set statements after modules like kivy.core.window are imported. It is a difficult problem to spot for a new user, and should be noted in the documentation.