Implement mmap() usage for raw texture data
This is an experiment to use mmap() for uploading texture data into
GPU. It can work only with raw texture data. It does work as intended,
but performance as not as big as wished, maybe to a synchronization
primitive or so.

Usage: Manually load a core image png/jpeg, and save with .rgb
extension (you don’t care if it’s rgb or rgba.). Then use the .rgb as
intended. To fully exploit the mmap() approach, the original file must
not be closed right away after populate texture. So add keep_data=True
in your image widget. Ideally, we should have a way to close the Mmap()
object after few frames if we don’t want to keep the data. The default
kivy behavior is to load the data, upload to gnu, and free everything.
mmap() have no impact, as it should still load everything in memory
_right now_. Goal is to delay the loading as much as possible and do
others thing in the meantime.

Careful: code is Unix only, mmap need to be adapted for Windows.