doc: No longer packaging *all* files in source dir
This is somewhat subjective (and perhaps both options should be given)
but generally I don't want to package every single file in my source
project to go into the output folder. I've added an example which just
picks kv and png files - I think this is better than the current example
(using Tree) because it's easier to cut mine down to add all files than
it is to expand the existing example to cherry-pick.