Fix examples/ PEP8 errors.  Mostly white space.
This fixes a large number of white space errors in the examples/
directory.
When I upgraded Makefile to check the 'make style' on the examples
directory,
the output was hundreds of lines long.   Currently, the output is
managable
and consists solely of "lines too long", as kv lines cannot be split.
This
commit touches an obscene number of files in trivial ways.