examples/ - Update docstrings
Update or create docstrings from many of the examples.  This is part of
the gallery project, as the new docstrings will be the website visible
documentation for the example.   Some changes are to use the main
phrase of ‘the file xx.py’ to include related files on the gallery page.

These are all documentation changes only.