Document and fix audio example
This provides documentation for the audio example; prevents it from
crashing when the SoundLoader fails; adds an example of mixing; allows
the player to attempt alternate formats if they are provided; adds a
screenshot so the demo is included in the audio files.

The piano sound was cut and downsampled to 160K per @dessant’s
complaints.