behaviors.py; fix keyboard eliding characters bug.
The keyboard input would clump characters together in a continuously
growing stream instead of timing out after a seond.  That is, typing
’12’, pausing, then typing ’34’ would register as a single string
‘1234’.   This showed up in the widgets/compound_selection.py example
and possible in issues with keyboard recognition problems.  The issue
is that time.clock() returns a vaguely defined ‘time on processor’
whereas time.time() is the number of seconds.