sdl2: rework support for pause/resume on iOS / Android
The previous implementation wasn’t fully working as @rnix explained.
This approach use the event filter from SDL2, in order to respond to
the OS synchronously. Because of this, an additional _pause_loop
attribute is added to ignore all SDL2 event and stay in the mainloop
until we go out of the pause. This prevent any Clock to work during the
pause.

Tested with on_pause return True, False, on iPad 2 simulator and iPad 3
real device, and desktop (nothing happen, but changes doesn’t impact
normal behavior.)

Fixes #3202