Docs: New Linux install instructions for Virtualenv and SDL2
Virtualenv now defaults to using --no-site-packages and that's how it
should be used, contain all software in the virtualenv.
Also there's no need to do a lot of wierd things to the users system.
Recommend users to install the exact version of Cython they need
with virtualenv.
No need to keep instructions for ancient Ubuntu versions (11.10).