Maintain last selection if not allowing empty selection
Self-explanatory. There doesn't seem any good reason to always default to the first item if the user deselects the last item, in the case where allow_empty_selection is already False.

In particular, this means accidental double/multiple-selections (especially common when using a mouse rather than touch) become less of an issue, and the likely intention of the user is carried through.