Add option to mimic Spinner height in dropdown
A fairly arbitrary size 48 height is currently the default (where's this set, by the way? Certainly not in this file. I couldn't find it).

This patch adds an option `mimic_size` which would have each dropdown element's height set according to the current size of the Spinner. More visually pleasing IMO. User customization is unaffected, perhaps a note should be made that this overwrites any changes in SpinnerOption.

For consideration before accepting:-
mimic_size or mimic_height?
I only set item.height (based on self.height) as size_hint_x of the default SpinnerOption is 1, hence setting item.size would not make a difference and may be a bit unclear.