Added Spinner text property update when values are updated.
If Spinner's values property is updated, the Spinner text is not changing. This is the wrong behavior. Let's make some examples:

1. User creates the Spinner with the values ('Home', 'Work', 'Holidays') and without the text property set. Spinner will show no text when some options are available; there will be no default option shown - and it is bad for the user experience.
2. User creates the Spinner with the empty values property. Then the Spinner instance is updated with the list of ('Home', 'Work', 'Holidays') but Spinner wont show any default value - empty text string suggests that no options are available to choose from.
3. User creates the Spinner with the values set to ('Home', 'Work', 'Holidays') and text set to 'Home'. Then Spinner values are updated to ('New Home', 'Another Work', 'Next Holidays'). The text will still show 'Home' as selected value, so without the user interaction the selection will be wrong.

I have added the text_autoupdate property that indicates if the user want the Spinner to automatically set the text property to values[0].

If set to True and values is not empty, the Spinner will automatically set its text to values[0]. Else, it will leave the text value untouched.

I believe it will make Spinner widget funtionality and complexity better,