Bug fix: value of -1 anim_delay setting correctly stops an image animation
The documentation  in kivy.uix.Image.anim_delay states that you can set
an value of =-1 to cause an animation to be stopped.
kivy.core.image.Image._set_anim_delay() checks for this, however
kivy.core.image.Image.anim_reset() does not, incorrectly scheduling a
Clock event for -1 (same frame) leading to a “too much iteration”
error. This fix adds a check to anim_reset() to only schedule the Clock
event if anim_delay is >= 0.