New ColorProperty and rgba symbol in kivy lang always accessible.
This PR intend to improve a little bit the color support in Kivy.
Currently, the 0-1 notation give flexibility, but it's hard to read it.

ColorProperty can accept #rrggbbaa or O-1 value as before. It doesn't accept 0-255.


A new rgba function in kivy.utils is added and exposed in kivy.lang.
rgba function accept 0-255 range value and hex value. rgba outputs always 4 values in 0-1 range, suitable for ColorProperty or ListProperty as today.