Store kivy_home_dir as a unicode string in python 2.
For Python 3 there is no change. For python 2 and users without
special characters, there is no change (because the string is easily
decoded with ascii codec anyway). The only change is for users that
have special characters in their username.

See https://github.com/kivy/kivy/issues/3912#issuecomment-229679674 for rationale and issue #3912 for reproduction.
Fixes #3912.