Use consistent Parser headers for logs instead of misleading "WARNING" title coming from nowhere
Logger.warning itself is a warning, no need to repeat it twice. User wants to know which part of Kivy raises such warning.

My first idea was to use "Lang" instead of "Parser" header, but for that we should use it everywhere in lang package. Ill be glad to hear what you think about it.