add an "auto" parameter to postproc:calibration
This allows to manage touch calibration non-fullscreen windows in a
relatively automatic way, which reacts to window's size and position
change.

Assumes that the device's input area exactly covers the display area.

This still require the configuration to indicate the display/input area size,
because we don't have a portable way of detecting that, but libraries
like screeninfo could be used for that.