Issue #5956: Fix casts in texture.blit_buffer for ushort and uint types.
Issue: When trying to use ushort and uint types, the cast to short and int
ctypes fail and result in an exception.

Fix: Do a cast to unsigned short and unsigned int instead.

Testing: Tested with the same diff on the release version 1.10.1. Tested
blitting a buffer of type int16, uint16, int32 and uint32. All types now
work as expected with no exceptions.