add magical load_kv method
This method looks up the name of the module that calls it, and load the
kv file with the same name but kv extension.

example usage:

in src/widgets/thing.py

    import load_kv
    from kivy.uix.widget import Widget

    class Thing(Widget):
        pass

    load_kv()

in src/widgets/thing.kv

    <Thing>:
        Label:
            text: 'magic!'

This also adds a packaging tool that will inline all such loaded kvlang
in the python code.

    python kivy/tools/packaging/bundle_kv.py -y

after this is called, all python code that contains the load_kv() call
will be edited *in place* to replace that call with
Builder.load_string("content of the kv file"), and the kv file *will be
deleted*, don't do this if you don't have uncommited code.