GL: Prevent enabling vertex attribute that are not in the shader
By glsl compiler optimizer or user error, it is possible that some 
attributes are not declared/available in the shader, but we are still 
enabling and assign them.

Let's say in 3D, you have v_pos, v_normal, v_tc0. And in the shader, 
only v_pos/v_normal are declared and used. All the call concerning v_tc0 
will fail.

We cannot use glGetError because it would fully sync the internal GL 
pipeline.
However, we can check that it returned -1, and avoid the call.