shrinks main content to 50%
It has been shown many times that humans can read more easily if
line lengths are bound. Optimal would be <= 80 characters per line,
however this is not practica for a documentation. My implementation
contains about 140 cpl, while the previous setting allowed for up
to 220.

It just makes it much more comfortable and easier to read.