Supports live resizing (sdl2 only)
Today, when the window is resized content is black or just partially
drawed. This is because poll() is blocked until we release the mouse.
But we can catch the resize event in the event filter. So the idea is
to catch and force the window resize, and force drawing.
Works like a charm.
Tested on OSX 10.14 with SDL 2.0.5 and retina screen.

Credits goes to @arafangion for the initial code.

Fixes #2844