Fixes SDL2 Mousewheel reading
We were using the wrong struct, therefore the wrong information was reported, and the code involved detection was also flawed (the if condition was badly written).

This commit uses the right struct, and fix the action name. Event works as intented, and scrollview left/right works on OSX now.

Fixes #2486