Fix scrolling with ctrl+cursor_up, ctrl+cursor_down
As the TextInput.do_cursor_movement documentation states the
"ctrl+cursor_up" and "ctrl+cursor_down" combinations should scroll text
one line up or down respectivelly without changing the cursor position.
Also the text should be scrolled both when the cursor is inside the
viewport and outside of it.
These 2 conditions were broken and are fixed by this commit.